Relocate page renderer into logic folder
