+++ sashimi-webapp/src/logic/renderer/index.js
